+++ myuw/tests.py
@@ -53,4 +53,3 @@ from myuw.test.views.test_api import TestDispatchErrorCases
 from myuw.test.views.textbooks import TestTextbook
 from myuw.test.speed import TestPageSpeeds
 from myuw.test.dao.visual_schedule import TestVisualSchedule
-from myuw.test.views.notice_admin import TestNoticeAdmin
